chore(nextjs): Extract internal nav function code into a hook
