+++ .changeset/small-frogs-march.md
@@ -1,5 +1,5 @@
 ---
-'@clerk/nextjs': minor
+'@clerk/nextjs': patch
 ---
 
 Utilize an awaitable replace function internally to avoid race conditions when using `router.replace`.
